requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
aiosqlite==0.19.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic==2.6.3
python-jose==3.3.0
langgraph==0.0.37
playwright==1.41.2
playwright-stealth==1.0.6
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.21.3
jinja2==3.1.3
email-validator==2.1.1
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
setuptools<81
